Get the Project to Build and Run.
Step 1 of the Code Lab is not building, as is. Making these changes, I got it to build and run on the Emlator API 26.
Moving the versions of the dependencies into the app/build.gradle file now gives you updates when a new version is available, so it's better than having it in the top level build.gradle, for a single module project.